filterList id removed / delegated event fix
The dropdown click functions were binding and unbinding to the body.
When a dropdown was closed, the click was unbound from the body, which
was also removing the delegated events click action. This caused app
events to stop firing.
Bound the drop down action to the appView (since it is used abundantly
below that level).

Also removed the id from the filterListTemplate to prevent duplicate
id's from being generated.