end cat static id fix / version
Static shows the page based on the id passed in the route obj. The id
for the sub cat, and the leaf cat are the same so after the first sub
cat in a root was viewed, only that cat was shown when other cat were
clicked.
The "end" param (end of the hash) for the leaf cat is now appended to
it's id to generate a unique id for the leaf.