Root splash template w/ list links to all
To allow team category list items to link directly to "all" page,
bypassing the type page which lists all, shirts, jerseys, etc., a new
template was created. The template translates w/ a list that places the
all link on the item, instead of the default.
This allows for the type landing page to be preserved for use with
direct links, but also for the user to land on the all page directly
from the list of teams.