Prod page ORDER SOON / qty adjustments
Set order soon to display once a size is selected.
The qty will be based on the variation's inventory instead of the group
inventory.
For some reason a size w/ 0 inventory was being added in the test
product, so handling for that was added as well: Qty select list and add
to cart button are disabled and a message to the user indicating size is
not available.