cart modal reloads checkout / version
When the cart was opened from checkout and changes were made, simply
closing it did not refresh the items in checkout.
The "close" cart model action was changed to remove checkout and show it
again in the event that the "checkout" or modal close buttons were used,
and the document location was already checkout.