qty select/cont shopping styles
Changed width of wrapper on select list to better hide default styling.
Added margin on cart qty section to space away from other lines (the
select is rendered differently in Safari, and is too close to the line
above).
Second attempt at fixing continue shopping button so it's not rendered
on two lines (except in FF).