Filtered search category loading fixes
Category/product lists were not being shown because template id's were
not being declared for most of the alias. They have been added for all
but promo, which needs reworking.
The dataset also needed to be adjusted for the subcat, and root aliases.

Also adjusted the class that was combining w/ the form to make the
selector used to trigger the initial form submission.