cart qty error fix / zip fix / chkout bttn default
Error where shown when qty added to cart exceeds qty available was
showing an undefined value because one of the vars was not updated when
the new appProductGet inventory method was added.
Updated qty and zip handling functions that were not moved to newinit in
update.
Problem w/ newShowContent taking checkout as an argument so custom
checkout button in cart was set to default until resolved.