remove link form vert nav hover elements
The element that triggers the hover out menus in the red vert nav ribbon
had a link on it.
This was problematic for touch devices because the link was triggered on
touch (simulated click). The hover out menu would start to show, but
then the page transition would take place and the user wouldn't be able
to select a team from within the menu shown.
The link was moved to the top of the first column in the menu that is
shown, in all caps text so it will stand out.
This prevents unreliable touch device detection from causing a bad user
experience.