prod list show/hide filters -> css animation / version
The js function that was animating the height used a fixed height which
had to allow for all filters to be open. If all of the filters were not
open, there was a large white space between the filters and the product.
This wasn't user friendly.
CSS animation allows for the container's content to set the height. As
filters are expanded, so is the container.