header banner reinitialize / version
This banner will be de-initialized, and then re-initialized on change of
the screen width.
This will allow the sizing to be updated when going from portrait to
landscape resolutions.
The homepage main banner is also set to do this, but the code is
commented out in the event because it would only be needed when a large
screen size change that is not likely to happen occurs (ie: changing
from phone resolution to desktop). However it can be quickly added if it
is found to be needed.