update brands json w/ multiple codes / version
New brands were added, and several on both the new and old set had
multiple vendor codes.
The a category object for each code was being generated resulting in
different categories w/ the same name and link.
Created an OR in the base filter for each of these categories so
products for both codes would show in a single category for that
brand/vendor.
The json generator will need to be altered to do this automatically.