filters on search results finished
Made new json map w/ only attribs for filters to be used w/ search.
Adjusted appendHash so get this json, grab the data for each attrib, and
add it to the dataset.
Created template for filter form to be transmogrified into the two
search templates.
Adjusted style for filter facet container to compensate for template
filter form.